change-database-sqlite-to-postgress

+++ db/migrate/20200325070551_add_fields_to_companies.rb
@@ -9,8 +9,8 @@ class AddFieldsToCompanies < ActiveRecord::Migration[6.0]
     add_column :companies, :city, :string
     add_column :companies, :state, :string
     add_column :companies, :country, :string
-    add_column :companies, :pincode, :integer
-    add_column :companies, :contact_no, :integer
+    add_column :companies, :pincode, :bigint
+    add_column :companies, :contact_no, :bigint
     add_column :companies, :website, :string
     add_column :companies, :status, :integer, default: 0
     add_column :companies, :facebook, :string
