+++ db/migrate/20200327113043_add_fields_to_users.rb
@@ -5,7 +5,7 @@ class AddFieldsToUsers < ActiveRecord::Migration[6.0]
   def change
     add_column :users, :first_name, :string
     add_column :users, :last_name, :string
-    add_column :users, :mobile, :integer
+    add_column :users, :mobile, :bigint
     add_column :users, :gender, :integer
     add_column :users, :active, :integer
   end
